+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.3.5.0 -- Unreleased
 
-* New built-in functions: `\deinterleave`, `\maximumBy`, `\minimumBy`.
+* New built-in functions: `\deinterleave`, `\maximumBy`, `\minimumBy`, `\unitVec2D`.
+* New particle: `\lengthWhile`.
 
 ### Breaking changes